Chess-Tutor-main/requirements.txt
Flask
openai==0.28.1
streamlit
easyocr
Pillow
python-dotenv
